split stream structure
